# Proof of concept: Transpile in workers