--- a/lib/babel/bundle.js
+++ b/lib/babel/bundle.js
@@ -19,10 +19,12 @@ function create(babelConfig) {
 	}
 
 	var start = Date.now();
+	var srcFile = path.join(cacheDir, 'faux-src.js');
 	var testFile = path.join(cacheDir, 'faux-test.js');
 	var stubFile = path.join(cacheDir, 'bundle-stub.js');
 
-	fs.writeFileSync(testFile, '');
+	fs.writeFileSync(srcFile, '');
+	fs.writeFileSync(testFile, 'require("./faux-src")');
 	fs.writeFileSync(stubFile, [
 		'var babelWorker = require(' + JSON.stringify(WORKER) + ')',
 		'babelWorker.runTest(' + JSON.stringify(testFile) + ', ' + JSON.stringify(babelConfig) + ')'