--- a/lib/babel/cache-worker.js
+++ b/lib/babel/cache-worker.js
@@ -1,34 +1,24 @@
 'use strict';
 var fs = require('fs');
 var path = require('path');
-
-var babel = require('babel-core');
 var findCacheDir = require('find-cache-dir');
 var md5hex = require('md5-hex');
-var objectAssign = require('object-assign');
 var pirates = require('pirates');
 
-function runTest(testPath, babelConfig) {
-	var options = objectAssign({
-		filename: testPath,
-		sourceMaps: true,
-		ast: false,
-		babelrc: false
-	}, babelConfig);
-
-	var start = Date.now();
-	var code = fs.readFileSync(testPath, 'utf8');
-	var result = babel.transform(code, options);
-	console.error('transformation took', Date.now() - start);
-
+function runTest(opts) {
+	var testPath = opts.testPath;
+	var code = fs.readFileSync(testPath);
 	var hash = md5hex(code);
+
 	var cacheDir = findCacheDir({name: 'ava', create: true});
 	var cachedFile = path.join(cacheDir, hash + '.js');
-	fs.writeFileSync(cachedFile, result.code);
+	if (!fs.existsSync(cachedFile)) {
+		return opts.loadWorker().runTest(testPath, opts.babelConfig);
+	}
 
 	var revert = pirates.addHook(
 		function () {
-			return result.code;
+			return fs.readFileSync(cachedFile, 'utf8');
 		},
 		{
 			exts: [path.extname(testPath)],