--- a/lib/babel/cache-worker.js
+++ b/lib/babel/cache-worker.js
@@ -1,9 +1,9 @@
 'use strict';
 var fs = require('fs');
 var path = require('path');
+var js = require('default-require-extensions/js');
 var findCacheDir = require('find-cache-dir');
 var md5hex = require('md5-hex');
-var pirates = require('pirates');
 
 function runTest(opts) {
 	var testPath = opts.testPath;
@@ -16,18 +16,22 @@ function runTest(opts) {
 		return opts.loadWorker().runTest(testPath, opts.babelConfig);
 	}
 
-	var revert = pirates.addHook(
-		function () {
-			return fs.readFileSync(cachedFile, 'utf8');
-		},
-		{
-			exts: [path.extname(testPath)],
-			matcher: function (filename) {
-				return filename === testPath;
-			}
+	// Guard against babel-core/register overriding the default extension.
+	var oldExtension = require.extensions['.js'];
+	require.extensions['.js'] = function (module, filename) {
+		if (filename !== opts.testPath) {
+			return oldExtension(module, filename);
 		}
-	);
+
+		require.extensions['.js'] = oldExtension;
+		var oldCompile = module._compile;
+		module._compile = function () {
+			module._compile = oldCompile;
+			module._compile(fs.readFileSync(cachedFile, 'utf8'), testPath);
+		};
+
+		js(module, testPath);
+	};
 	require(testPath); // eslint-disable-line import/no-dynamic-require
-	revert();
 }
 exports.runTest = runTest;