--- a/lib/babel/cache-worker.js
+++ b/lib/babel/cache-worker.js
@@ -5,6 +5,8 @@ var js = require('default-require-extensions/js');
 var findCacheDir = require('find-cache-dir');
 var md5hex = require('md5-hex');
 
+var register = require('./register');
+
 function runTest(opts) {
 	var testPath = opts.testPath;
 	var code = fs.readFileSync(testPath);
@@ -16,14 +18,9 @@ function runTest(opts) {
 		return opts.loadWorker().runTest(testPath, opts.babelConfig);
 	}
 
-	// Guard against babel-core/register overriding the default extension.
-	var oldExtension = require.extensions['.js'];
-	require.extensions['.js'] = function (module, filename) {
-		if (filename !== opts.testPath) {
-			return oldExtension(module, filename);
-		}
+	require.extensions['.js'] = function (module) {
+		register.install();
 
-		require.extensions['.js'] = oldExtension;
 		var oldCompile = module._compile;
 		module._compile = function () {
 			module._compile = oldCompile;