--- a/lib/babel/config.js
+++ b/lib/babel/config.js
@@ -22,7 +22,7 @@ var DEFAULT_PRESETS = (function () {
 
 var REQUIRED_PLUGINS = [
 	'babel-plugin-ava-throws-helper',
-	// REWRITE_RUNTIME, // FIXME Not yet compatible with cached-module-loader
+	REWRITE_RUNTIME,
 	'babel-plugin-transform-runtime'
 ];