--- a/lib/babel/worker.js
+++ b/lib/babel/worker.js
@@ -3,10 +3,10 @@ var fs = require('fs');
 var path = require('path');
 
 var babel = require('babel-core');
+var js = require('default-require-extensions/js');
 var findCacheDir = require('find-cache-dir');
 var md5hex = require('md5-hex');
 var objectAssign = require('object-assign');
-var pirates = require('pirates');
 
 function runTest(testPath, babelConfig) {
 	var options = objectAssign({
@@ -26,18 +26,22 @@ function runTest(testPath, babelConfig) {
 	var cachedFile = path.join(cacheDir, hash + '.js');
 	fs.writeFileSync(cachedFile, result.code);
 
-	var revert = pirates.addHook(
-		function () {
-			return result.code;
-		},
-		{
-			exts: [path.extname(testPath)],
-			matcher: function (filename) {
-				return filename === testPath;
-			}
+	// Guard against babel-core/register overriding the default extension.
+	var oldExtension = require.extensions['.js'];
+	require.extensions['.js'] = function (module, filename) {
+		if (filename !== testPath) {
+			return oldExtension(module, filename);
 		}
-	);
+
+		require.extensions['.js'] = oldExtension;
+		var oldCompile = module._compile;
+		module._compile = function () {
+			module._compile = oldCompile;
+			module._compile(result.code, testPath);
+		};
+
+		js(module, testPath);
+	};
 	require(testPath); // eslint-disable-line import/no-dynamic-require
-	revert();
 }
 exports.runTest = runTest;