--- a/lib/babel/worker.js
+++ b/lib/babel/worker.js
@@ -8,6 +8,8 @@ var findCacheDir = require('find-cache-dir');
 var md5hex = require('md5-hex');
 var objectAssign = require('object-assign');
 
+var register = require('./register');
+
 function runTest(testPath, babelConfig) {
 	var options = objectAssign({
 		filename: testPath,
@@ -26,14 +28,9 @@ function runTest(testPath, babelConfig) {
 	var cachedFile = path.join(cacheDir, hash + '.js');
 	fs.writeFileSync(cachedFile, result.code);
 
-	// Guard against babel-core/register overriding the default extension.
-	var oldExtension = require.extensions['.js'];
-	require.extensions['.js'] = function (module, filename) {
-		if (filename !== testPath) {
-			return oldExtension(module, filename);
-		}
+	require.extensions['.js'] = function (module) {
+		register.install();
 
-		require.extensions['.js'] = oldExtension;
 		var oldCompile = module._compile;
 		module._compile = function () {
 			module._compile = oldCompile;