--- a/lib/cli.js
+++ b/lib/cli.js
@@ -15,7 +15,7 @@ var miniReporter = require('./reporters/mini');
 var tapReporter = require('./reporters/tap');
 var Logger = require('./logger');
 var Watcher = require('./watcher');
-var babelConfig = require('./babel-config');
+var babelConfig = require('./babel/config');
 
 // Bluebird specific
 Promise.longStackTraces();