--- a/lib/logger.js
+++ b/lib/logger.js
@@ -98,6 +98,6 @@ Logger.prototype.exit = function (code) {
 
 	// timeout required to correctly flush IO on Node.js 0.10 on Windows
 	setTimeout(function () {
-		process.exit(code); // eslint-disable-line xo/no-process-exit
+		process.exit(code); // eslint-disable-line unicorn/no-process-exit
 	}, process.env.AVA_APPVEYOR ? 500 : 0);
 };