--- a/lib/process-adapter.js
+++ b/lib/process-adapter.js
@@ -16,7 +16,7 @@ if (!isForked) {
 	console.log();
 	console.error('Test files must be run with the AVA CLI:\n\n    ' + chalk.grey.dim('$') + ' ' + chalk.cyan('ava ' + fp) + '\n');
 
-	process.exit(1); // eslint-disable-line xo/no-process-exit
+	process.exit(1); // eslint-disable-line unicorn/no-process-exit
 }
 
 exports.send = function (name, data) {