--- a/lib/process-adapter.js
+++ b/lib/process-adapter.js
@@ -1,9 +1,6 @@
 'use strict';
-var fs = require('fs');
 var path = require('path');
 var chalk = require('chalk');
-var sourceMapSupport = require('source-map-support');
-var installPrecompiler = require('require-precompiled');
 
 var debug = require('debug')('ava');
 
@@ -61,37 +58,6 @@ if (debug.enabled) {
 
 	require('time-require'); // eslint-disable-line import/no-unassigned-import
 }
-
-var sourceMapCache = Object.create(null);
-var cacheDir = opts.cacheDir;
-
-exports.installSourceMapSupport = function () {
-	sourceMapSupport.install({
-		environment: 'node',
-		handleUncaughtExceptions: false,
-		retrieveSourceMap: function (source) {
-			if (sourceMapCache[source]) {
-				return {
-					url: source,
-					map: fs.readFileSync(sourceMapCache[source], 'utf8')
-				};
-			}
-		}
-	});
-};
-
-exports.installPrecompilerHook = function () {
-	installPrecompiler(function (filename) {
-		var precompiled = opts.precompiled[filename];
-
-		if (precompiled) {
-			sourceMapCache[filename] = path.join(cacheDir, precompiled + '.js.map');
-			return fs.readFileSync(path.join(cacheDir, precompiled + '.js'), 'utf8');
-		}
-
-		return null;
-	});
-};
 
 exports.installDependencyTracking = function (dependencies, testPath) {
 	Object.keys(require.extensions).forEach(function (ext) {