--- a/lib/reporters/mini.js
+++ b/lib/reporters/mini.js
@@ -159,12 +159,8 @@ MiniReporter.prototype.finish = function (runStatus) {
 		status += '\n  ' + colors.error(runStatus.previousFailCount, 'previous', plur('failure', runStatus.previousFailCount), 'in test files that were not rerun');
 	}
 
-	var i = 0;
-
 	if (this.knownFailureCount > 0) {
 		runStatus.knownFailures.forEach(function (test) {
-			i++;
-
 			var title = test.title;
 
 			status += '\n\n   ' + colors.title(title);
@@ -178,8 +174,6 @@ MiniReporter.prototype.finish = function (runStatus) {
 			if (!test.error || !test.error.message) {
 				return;
 			}
-
-			i++;
 
 			var title = test.error ? test.title : 'Unhandled Error';
 			var description;
@@ -209,8 +203,6 @@ MiniReporter.prototype.finish = function (runStatus) {
 			if (err.title) {
 				return;
 			}
-
-			i++;
 
 			if (err.type === 'exception' && err.name === 'AvaError') {
 				status += '\n\n  ' + colors.error(cross + ' ' + err.message);