--- a/lib/test-worker.js
+++ b/lib/test-worker.js
@@ -15,8 +15,6 @@ Promise.longStackTraces();
 
 (opts.require || []).forEach(require);
 
-process.installSourceMapSupport();
-
 var currentlyUnhandled = require('currently-unhandled')();
 var serializeError = require('./serialize-error');
 var send = process.send;
@@ -24,8 +22,6 @@ var throwsHelper = require('./throws-helper');
 
 // check if test files required ava and show error, when they didn't
 exports.avaRequired = false;
-
-process.installPrecompilerHook();
 
 var dependencies = [];
 process.installDependencyTracking(dependencies, testPath);