--- a/lib/test-worker.js
+++ b/lib/test-worker.js
@@ -16,6 +16,7 @@ Promise.longStackTraces();
 (opts.require || []).forEach(require);
 
 var currentlyUnhandled = require('currently-unhandled')();
+var babelWorker = require('./babel/worker');
 var serializeError = require('./serialize-error');
 var send = process.send;
 var throwsHelper = require('./throws-helper');
@@ -26,7 +27,7 @@ exports.avaRequired = false;
 var dependencies = [];
 process.installDependencyTracking(dependencies, testPath);
 
-require(testPath); // eslint-disable-line import/no-dynamic-require
+babelWorker.runTest(testPath, opts.babel);
 
 process.on('unhandledRejection', throwsHelper);