--- a/lib/test-worker.js
+++ b/lib/test-worker.js
@@ -15,8 +15,9 @@ Promise.longStackTraces();
 
 (opts.require || []).forEach(require);
 
+var fs = require('fs');
 var currentlyUnhandled = require('currently-unhandled')();
-var babelWorker = require('./babel/worker');
+var babelBundle = require('./babel/bundle');
 var serializeError = require('./serialize-error');
 var send = process.send;
 var throwsHelper = require('./throws-helper');
@@ -27,6 +28,10 @@ exports.avaRequired = false;
 var dependencies = [];
 process.installDependencyTracking(dependencies, testPath);
 
+var babelWorker = babelBundle.loadWorker({
+	cachedData: fs.readFileSync(opts.bundle.cachedDataFile),
+	code: fs.readFileSync(opts.bundle.codeFile)
+});
 babelWorker.runTest(testPath, opts.babel);
 
 process.on('unhandledRejection', throwsHelper);