--- a/lib/test-worker.js
+++ b/lib/test-worker.js
@@ -13,7 +13,8 @@ var Promise = require('bluebird');
 // Bluebird specific
 Promise.longStackTraces();
 
-(opts.require || []).forEach(require);
+// Ignore, assume this only installs babel-register which is now built-in.
+// (opts.require || []).forEach(require);
 
 var fs = require('fs');
 var currentlyUnhandled = require('currently-unhandled')();