… cli.js
… lib/babel-config.js
… package.json
… scripts/xo.js
… test/cli.js
… test/visual/lorem-ipsum.js
… test/visual/run-visual-tests.js
… test/watcher.js