@novemberborn 2017-01-17
This will be fleshed out further when implementing [RFC001](https://github.com/avajs/ava/blob/master/docs/specs/001%20-%20Improving%20language%20support.md).